Moved parser.py into own package dir.
